+++ example/typescript_react/src/App.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import {
-  initialize,
   h,
   SExp,
   CLVMType,
@@ -22,13 +21,6 @@ function App() {
   // eslint-disable-next-line max-len
   const [prg, setPrg] = React.useState('0xff01ff01ffffffa09dc16b766b557d0d8d94fe1ee636245b4417a46cd53bd4e70c26a62dc698d406ffff02ffff01ff02ffff01ff02ffff03ff0bffff01ff02ffff03ffff09ff05ffff1dff0bffff1effff0bff0bffff02ff06ffff04ff02ffff04ff17ff8080808080808080ffff01ff02ff17ff2f80ffff01ff088080ff0180ffff01ff04ffff04ff04ffff04ff05ffff04fffe84016b6b7fff80808080fffe820db78080fe81ffffff04ffff01ff32ff02ffff03ffff07ff0580ffff01ff0bffff0102ffff02ff06ffff04ff02ffff04ff09ff80808080ffff02ff06ffff04ff02ffff04ff0dff8080808080ffff01ff0bffff0101fe6f80ff0180fe3e80ffff04ffff01b0a282b4b0117a8d04906835fa4fa0e13d3fbd1dd61899ebdf5157977611d1bae52f2ea97cbc3916466b1c9176d30a9030fe3f80ff85727e956731ffff80ffff01ffff33ffa05597ef68eaf171a6303995ecbb14fdbf2c24300b625bfbc886ea68270424661dff5880ffff33ffa07a7a9cb053b9e7086ddbb789a4a1abc646a06627d372eca59368bf90c15028bfff85727b9a765980ffff34ff8402faf08080ffff3cffa0d71f4c45af09583209498dbb9974bbda21b859fac0bf3348337ed33a2ba5c3838080ff808080808080');
   const [env, setEnv] = React.useState('0x80');
-  
-  React.useEffect(() => {
-    // Load blsjs.wasm and clvm_wasm_bg.wasm
-    initialize().catch((e: unknown) => {
-      setOutput(e instanceof Error ? e.message : JSON.stringify(e));
-    });
-  }, []);
   
   const onInputChange = React.useCallback((e: React.ChangeEvent<HTMLInputElement>) => {
     const inputEl = e.currentTarget as HTMLInputElement;