+++ example/typescript_react/vite.config.ts
@@ -9,34 +9,15 @@ export default defineConfig(() => {
     build: {
       outDir: '.dist',
     },
-    resolve: {
-      /*
-        `resolve.preserveSymlinks: true` is only required when `clvm` is specified
-        as a link dependency.
-          "dependencies": {
-            "clvm": "link:../../.dist/npm", // <--- This one
-            "react": "^18.3.1",
-            "react-dom": "^18.3.1"
-          },
-        Usually you don't need this option to be set without
-        a link dependency like:
-          "dependencies": {
-            "clvm": "3.0.0",
-            "react": "^18.3.1",
-           "react-dom": "^18.3.1"
-          },
-       */
-      // preserveSymlinks: true,
-    },
     server: {
       port: SERVER_PORT,
       proxy: {
-        '^/(blsjs|clvm_wasm_bg).wasm': {
+        '^.*(?<!/public/assets/)(blsjs|clvm_wasm_bg).wasm': {
           target: {
             host: 'localhost',
             port: SERVER_PORT,
           },
-          rewrite: (path) => `/public/assets/${path.replace(/^\//, '')}`,
+          rewrite: (path) => `/public/assets/${path.replace(/.*((blsjs|clvm_wasm_bg).wasm)$/, '$1')}`,
         },
       },
     },